pyproject.toml
[project]
name = "backend-of-party-qa-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pyjwt>=2.10.1",
    "sqlmodel>=0.0.24",
]
